Update: 강화 장비 스텟 반영

+++ Source/ARPG_Hunter/Data/EquipmentUpgradeData.h
@@ -40,7 +40,7 @@ public:
 	UPROPERTY(EditAnywhere)
 	TArray<FUpgradeIngredient> Ingredients;
 	UPROPERTY(EditAnywhere)
-	TMap<ECharacterStatType, uint16> StatPerStep;
+	TMap<ECharacterStatType, uint32> StatPerStep;
 public:
 	static FName GetUpgradeID(uint8 _rank, uint8 _grade, EEquipmentType _type);
 };
